More refactors to remove java-syntax dependency from core
