+++ TODO.md
@@ -66,7 +66,14 @@ If the data plugin is removed from the build then
 
 ```
 Projection source failed: java.lang.NullPointerException: Cannot invoke "mx.sugus.braid.core.plugin.SmithyGeneratorPlugin.provides()" because "plugin" is null
+
 ```
+
+### Data Codgen
+
+* The package is duplicated, we need to find a place for this,
+  probably returning a CompilationUnit instead of a generic
+  TypeSyntax.
 
 ----
 ## Done
