+++ braid-core/src/main/java/mx/sugus/braid/core/BraidCodegenPlugin.java
@@ -1,23 +1,17 @@
 package mx.sugus.braid.core;
 
-import java.util.Objects;
-import java.util.function.Function;
 import mx.sugus.braid.core.plugin.ClassPathPluginLoader;
 import mx.sugus.braid.core.plugin.CodegenModule;
 import mx.sugus.braid.core.plugin.ComposedPluginLoader;
 import mx.sugus.braid.core.plugin.DefaultBaseModuleConfig;
 import mx.sugus.braid.core.plugin.PluginLoader;
 import mx.sugus.braid.core.plugin.SpiPluginLoader;
-import mx.sugus.braid.core.symbol.BraidSymbolProvider;
 import software.amazon.smithy.build.PluginContext;
 import software.amazon.smithy.build.SmithyBuildPlugin;
-import software.amazon.smithy.codegen.core.ReservedWords;
-import software.amazon.smithy.codegen.core.ReservedWordsBuilder;
 import software.amazon.smithy.codegen.core.SymbolProvider;
 import software.amazon.smithy.model.Model;
 
 public final class BraidCodegenPlugin implements SmithyBuildPlugin {
-    public static final ReservedWords RESERVED_WORDS = buildReservedWords();
 
     @Override
     public String getName() {
@@ -29,9 +23,8 @@ public final class BraidCodegenPlugin implements SmithyBuildPlugin {
         var settingsNode = context.getSettings();
         var settings = JavaCodegenSettings.from(settingsNode);
         var module = new CodegenModule(DefaultBaseModuleConfig.buildDependants(pluginLoader(), settings.settingsNode()));
-        var model = module.earlyPreprocessModel(context.getModel());
         BraidCodegenDirector.builder()
-                            .model(model)
+                            .model(context.getModel())
                             .settings(settings)
                             .fileManifest(context.getFileManifest())
                             .module(module)
@@ -45,19 +38,10 @@ public final class BraidCodegenPlugin implements SmithyBuildPlugin {
     }
 
     private static SymbolProvider createSymbolProvider(Model model, JavaCodegenSettings settings) {
-        var escaper = RESERVED_WORDS;
-        var shapeToJavaName = new ShapeToJavaName(model, escaper, settings.packageName());
-        var shapeToJavaType = new ShapeToJavaType(shapeToJavaName, model);
-        return new BraidSymbolProvider(model, shapeToJavaName, shapeToJavaType, settings.packageName());
-    }
-
-    private static ReservedWords buildReservedWords() {
-        return
-            new ReservedWordsBuilder()
-                .loadWords(Objects.requireNonNull(BraidCodegenPlugin.class.getResource("java-reserved-words.txt")),
-                           Function.identity())
-                .loadWords(Objects.requireNonNull(BraidCodegenPlugin.class.getResource("java-system-type-names.txt")),
-                           Function.identity())
-                .build();
+        return shape -> {
+            throw new UnsupportedOperationException(
+                "shape to symbol not supported for shape type: " + shape.getType()
+                + ", shape id: " + shape.getId());
+        };
     }
 }
