+++ braid-data-plugin/src/main/java/mx/sugus/braid/plugins/data/DataPlugin.java
@@ -11,6 +11,7 @@ import mx.sugus.braid.plugins.data.producers.EnumJavaProducer;
 import mx.sugus.braid.plugins.data.producers.StructureInterfaceJavaProducer;
 import mx.sugus.braid.plugins.data.producers.StructureJavaProducer;
 import mx.sugus.braid.plugins.data.producers.UnionJavaProducer;
+import mx.sugus.braid.plugins.data.utils.DataSymbolProviderDecorator;
 import mx.sugus.braid.rt.util.annotations.Generated;
 import software.amazon.smithy.model.node.ObjectNode;
 
@@ -32,16 +33,13 @@ public final class DataPlugin implements SmithyGeneratorPlugin {
 
     @Override
     public CodegenModuleConfig moduleConfig(ObjectNode node) {
-        return newBaseConfig();
-    }
-
-    static CodegenModuleConfig newBaseConfig() {
         return CodegenModuleConfig
             .builder()
             .addProducer(new StructureJavaProducer())
             .addProducer(new StructureInterfaceJavaProducer())
             .addProducer(new EnumJavaProducer())
             .addProducer(new UnionJavaProducer())
+            .addSymbolProviderDecorator(new DataSymbolProviderDecorator(node.expectStringMember("package").getValue()))
             .build();
     }
