+++ braid-data-plugin/src/main/java/mx/sugus/braid/plugins/data/producers/BuilderAdderOverrides.java
@@ -13,7 +13,7 @@ import static mx.sugus.braid.plugins.data.producers.Utils.toMemberJavaName;
 import java.util.ArrayList;
 import java.util.List;
 import javax.lang.model.element.Modifier;
-import mx.sugus.braid.core.SymbolConstants;
+import mx.sugus.braid.plugins.data.utils.SymbolConstants;
 import mx.sugus.braid.core.plugin.ShapeCodegenState;
 import mx.sugus.braid.jsyntax.ClassName;
 import mx.sugus.braid.jsyntax.ClassSyntax;
