+++ braid-data-plugin/src/main/java/mx/sugus/braid/plugins/data/producers/EnumJavaProducer.java
@@ -3,7 +3,7 @@ package mx.sugus.braid.plugins.data.producers;
 import mx.sugus.braid.core.plugin.Identifier;
 import mx.sugus.braid.core.plugin.ShapeCodegenState;
 import mx.sugus.braid.core.plugin.ShapeProducerTask;
-import mx.sugus.braid.core.plugin.TypeSyntaxResult;
+import mx.sugus.braid.plugins.data.TypeSyntaxResult;
 import mx.sugus.braid.traits.JavaTrait;
 import software.amazon.smithy.model.shapes.ShapeType;
