+++ braid-data-plugin/src/main/java/mx/sugus/braid/plugins/data/producers/UnionDataBuilder.java
@@ -9,7 +9,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.function.Consumer;
 import javax.lang.model.element.Modifier;
-import mx.sugus.braid.core.SymbolConstants;
+import mx.sugus.braid.plugins.data.utils.SymbolConstants;
 import mx.sugus.braid.core.plugin.ShapeCodegenState;
 import mx.sugus.braid.core.util.Name;
 import mx.sugus.braid.jsyntax.CaseClause;
