+++ braid-data-plugin/src/main/java/mx/sugus/braid/plugins/data/utils/ShapeToJavaName.java
@@ -1,4 +1,4 @@
-package mx.sugus.braid.core;
+package mx.sugus.braid.plugins.data.utils;
 
 import mx.sugus.braid.core.util.Name;
 import software.amazon.smithy.codegen.core.ReservedWords;
