+++ braid-java-codegen-test/smithy-build.json
@@ -13,6 +13,7 @@
                 "mx.sugus.braid.plugins.data#FlattenInterfaceMembersPlugin": {
                 },
                 "mx.sugus.braid.plugins.data#DataPlugin": {
+                    "package": "mx.sugus.braid.test"
                 }
             }
         }
