+++ braid-serde-node-plugin/src/main/java/mx/sugus/braid/plugins/serde/node/InterfaceAddToNodeTransformer.java
@@ -3,7 +3,7 @@ package mx.sugus.braid.plugins.serde.node;
 import mx.sugus.braid.core.plugin.Identifier;
 import mx.sugus.braid.core.plugin.ShapeCodegenState;
 import mx.sugus.braid.core.plugin.ShapeTaskTransformer;
-import mx.sugus.braid.core.plugin.TypeSyntaxResult;
+import mx.sugus.braid.plugins.data.TypeSyntaxResult;
 import mx.sugus.braid.jsyntax.InterfaceSyntax;
 import mx.sugus.braid.plugins.data.producers.StructureInterfaceJavaProducer;
 import software.amazon.smithy.model.node.ToNode;
