+++ braid-syntax-plugin/src/main/java/mx/sugus/braid/plugins/syntax/InterfaceSyntaxAddAcceptVisitorTransformer.java
@@ -3,7 +3,7 @@ package mx.sugus.braid.plugins.syntax;
 import mx.sugus.braid.core.plugin.Identifier;
 import mx.sugus.braid.core.plugin.ShapeCodegenState;
 import mx.sugus.braid.core.plugin.ShapeTaskTransformer;
-import mx.sugus.braid.core.plugin.TypeSyntaxResult;
+import mx.sugus.braid.plugins.data.TypeSyntaxResult;
 import mx.sugus.braid.jsyntax.AbstractMethodSyntax;
 import mx.sugus.braid.jsyntax.ClassName;
 import mx.sugus.braid.jsyntax.InterfaceSyntax;
